improve: revert "feat: api getPasswordSettings, generate passwords in cli and api follow settings"
This reverts commit 73acf318d7ef2a70ddf2d436f507b9d3dc7671d9.
