+++ internal/access/credential.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/infrahq/infra/internal"
 	"github.com/infrahq/infra/internal/generate"
-	"github.com/infrahq/infra/internal/logging"
 	"github.com/infrahq/infra/internal/server/data"
 	"github.com/infrahq/infra/internal/server/models"
 	"github.com/infrahq/infra/internal/validate"
@@ -24,24 +23,9 @@ func CreateCredential(c *gin.Context, user models.Identity) (string, error) {
 		return "", HandleAuthErr(err, "user", "create", models.InfraAdminRole)
 	}
 
-	settings, err := data.GetSettings(db)
-	if err != nil {
-		return "", err
-	}
-
-GENERATE:
-	tmpPassword, err := generate.CryptoRandom(settings.LengthMin, generate.CharsetPassword)
+	tmpPassword, err := generate.CryptoRandom(12, generate.CharsetPassword)
 	if err != nil {
 		return "", fmt.Errorf("generate: %w", err)
-	}
-
-	err = checkPasswordRequirements(db, tmpPassword)
-	var validateError validate.Error
-	if err != nil {
-		if errors.As(err, &validateError) {
-			goto GENERATE
-		}
-		return "", err
 	}
 
 	hash, err := bcrypt.GenerateFromPassword([]byte(tmpPassword), bcrypt.DefaultCost)
@@ -139,33 +123,27 @@ func hasMinimumCount(min int, password string, check func(rune) bool) bool {
 func checkPasswordRequirements(db *gorm.DB, password string) error {
 	settings, err := data.GetSettings(db)
 	if err != nil {
-		logging.Debugf("error while getting settings")
 		return err
 	}
 	errs := make(validate.Error)
 
 	if !hasMinimumCount(settings.LowercaseMin, password, unicode.IsLower) {
-		logging.Errorf("error generating password: not enough lower case letters")
 		errs["password"] = append(errs["password"], fmt.Sprintf("needs minimum %d lower case letters", settings.LowercaseMin))
 	}
 
 	if !hasMinimumCount(settings.UppercaseMin, password, unicode.IsUpper) {
-		logging.Errorf("error generating password: not enough upper case letters")
 		errs["password"] = append(errs["password"], fmt.Sprintf("needs minimum %d upper case letters", settings.UppercaseMin))
 	}
 
 	if !hasMinimumCount(settings.NumberMin, password, unicode.IsDigit) {
-		logging.Errorf("error generating password: not enough numbers")
 		errs["password"] = append(errs["password"], fmt.Sprintf("needs minimum %d numbers", settings.NumberMin))
 	}
 
 	if !hasMinimumCount(settings.SymbolMin, password, isValidSymbol) {
-		logging.Errorf("error generating password: not enough symbols")
 		errs["password"] = append(errs["password"], fmt.Sprintf("needs minimum %d symbols", settings.SymbolMin))
 	}
 
 	if len(password) < settings.LengthMin {
-		logging.Errorf("error generating password: not long enough")
 		errs["password"] = append(errs["password"], fmt.Sprintf("needs minimum length of %d", settings.LengthMin))
 	}
